refactor: 의존성 관리, 의존성 주입 관심사 분리
